moved library to providers/solomon
