+++ ydb/library/yql/providers/common/ya.make
@@ -3,7 +3,6 @@ RECURSE(
     db_id_async_resolver
     http_gateway
     pushdown
-    solomon_accessor
     token_accessor
     ut_helpers
 )
