+++ ydb/library/yql/providers/solomon/solomon_accessor/client/ya.make
@@ -7,8 +7,8 @@ SRCS(
 PEERDIR(
     library/cpp/protobuf/interop
     ydb/library/yql/providers/common/http_gateway
-    ydb/library/yql/providers/common/solomon_accessor/grpc
     ydb/library/yql/providers/solomon/proto
+    ydb/library/yql/providers/solomon/solomon_accessor/grpc
     ydb/public/sdk/cpp/src/library/grpc/client
     yql/essentials/utils
 )
