+++ ydb/library/yql/providers/solomon/solomon_accessor/grpc/solomon_accessor_pb.proto
