+++ ydb/library/yql/providers/solomon/solomon_accessor/grpc/ya.make
